+++ src/components/chat/ChatEditForm.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react"
+import ChatManager from "../../modules/ChatManager"
+import ChatList from "./ChatList";
 
-export default class ChatForm extends Component {
-
+export default class ChatEdit extends Component {
     state = {
         text: "",
         // userId: ""
@@ -13,23 +14,28 @@ export default class ChatForm extends Component {
         this.setState(stateToChange)
     }
 
-    makeNewMessage = evt => {
+    updateExistingChat = evt => {
         evt.preventDefault()
 
-        const chats = {
-            text: this.state.chatMessage,
-            // userId: parseInt(this.state.userId)
+        const editedChat = {
+            id: this.props.match.params.chatId,
+            text: this.state.text
         }
 
-        this.props
-            .addChat(chats)
-            .then(() => this.props.history.push("/chats"));
+        this.props.updateExistingChat(editedChat)
+        .then(() => this.props.history.push("/chats"))
     }
 
+    componentDidMount() {
+        ChatManager.get(this.props.match.params.chatId)
+        .then(chat => {
+            this.setState({text: chat.text})
+        })
+    }
 
-render() {
-    return (
-        <React.Fragment>
+    render() {
+        return (
+            <React.Fragment>
             <form className="chatForm">
                 <div className="form-group">
                     <input
@@ -37,17 +43,17 @@ render() {
                     required
                     className="form-control"
                     onChange={this.handleFieldChange}
-                    id="chatMessage"
+                    id="text"
                     placeholder="chat message"
-                    // value={this.state.text}
+                    value={this.state.text}
                     />
                     <button
                         type="submit"
-                        onClick={this.makeNewMessage}
+                        onClick={this.updateExistingChat}
                         className="btn btn-primary">Submit</button>
                 </div>
             </form>
         </React.Fragment>
-    )
-}
+        )
+    }
 }
